Fixing warnings

+++ src/operators/radiationDiffusionFD/RadiationDiffusionModel.h
@@ -86,7 +86,7 @@ protected:
     double d_currentTime = 0.0;
 
     //! Shorthand for spatial dimension
-    size_t d_dim = -1;
+    size_t d_dim = (size_t) -1;
 
     //! Basic parameter database (with model-agnostic parameters)
     std::shared_ptr<AMP::Database> d_basic_db;
